[fix/debug] Debug+fix stratum

+++ libnet2/worker.h
@@ -44,7 +44,8 @@ public:
 
 struct NotifyData
 {
-    uint64_t version;
+//    uint64_t version;
+    int32_t version;
     uint256 previous_block;
     coind::data::MerkleLink merkle_link;
     std::vector<unsigned char> coinb1;
